fix: corrige nome do modulo para App e seus tipos
