+++ src/index.ts
@@ -4,7 +4,7 @@ import { StyleHelper } from './helpers/styleHelper'
 import { ViewportHelper } from './helpers/viewportHelper'
 
 // Modules services
-import { AppsService } from './modules/apps/AppsService'
+import { AppsService } from './modules/app/AppService'
 import { BlogCategoryService } from './modules/blog/category/BlogCategoryService'
 import { BlogPostService } from './modules/blog/post/BlogPostService'
 import { BrandService } from './modules/brand/BrandService'
